feat(dui): moves AvaloniaHwndHost to a separate proj
