+++ ConnectorRevit/ConnectorRevit/UI/Panel.xaml
@@ -2,9 +2,8 @@
   x:Class="Speckle.ConnectorRevit.Panel"
   xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
   xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-  xmlns:av="clr-namespace:DesktopUI2.Views;assembly=DesktopUI2"
   xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-  xmlns:local="clr-namespace:Speckle.ConnectorRevit"
+  xmlns:av="clr-namespace:AvaloniaHwndHost;assembly=AvaloniaHwndHost"
   xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
   Title="Page1"
   Width="400"
@@ -12,7 +11,7 @@
   mc:Ignorable="d">
 
   <Grid>
-    <local:AvaloniaHwndHost x:Name="AvaloniaHost" />
+    <av:AvaloniaHwndHost x:Name="AvaloniaHost" />
 
   </Grid>
 </Page>
