+++ DesktopUI2/DesktopUI2.WPF/DesktopUI2.WPF.csproj
@@ -5,7 +5,6 @@
     <TargetFramework>net5.0-windows</TargetFramework>
     <Nullable>enable</Nullable>
     <UseWPF>true</UseWPF>
-    <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
   </PropertyGroup>
 
   <ItemGroup>
@@ -15,6 +14,7 @@
   </ItemGroup>
 
   <ItemGroup>
+    <ProjectReference Include="..\AvaloniaHwndHost\AvaloniaHwndHost.csproj" />
     <ProjectReference Include="..\DesktopUI2\DesktopUI2.csproj" />
   </ItemGroup>
