+++ DesktopUI2/DesktopUI2.WPF/MainWindow.xaml
@@ -6,13 +6,14 @@
   xmlns:dui="clr-namespace:DesktopUI2.Views;assembly=DesktopUI2"
   xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
   xmlns:local="clr-namespace:DesktopUI2.WPF"
+  xmlns:av="clr-namespace:AvaloniaHwndHost;assembly=AvaloniaHwndHost"
   xmlns:wpf="clr-namespace:Avalonia.Win32.Interop.Wpf;assembly=Avalonia.Win32.Interop"
   Title="MainWindow"
   Width="400"
   Height="750"
   mc:Ignorable="d">
   <Grid>
-    <local:AvaloniaHwndHost x:Name="AvaloniaHost" />
+    <av:AvaloniaHwndHost x:Name="AvaloniaHost" />
     <!--<wpf:WpfAvaloniaHost>
       <dui:MainUserControl />
     </wpf:WpfAvaloniaHost>-->
