+++ DesktopUI2/DesktopUI2.sln
@@ -9,7 +9,9 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DesktopUI2.Launcher", "Desk
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Core", "..\Core\Core\Core.csproj", "{B4981AA3-6688-4AA0-B659-E8B54763D1A6}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DesktopUI2.WPF", "DesktopUI2.WPF\DesktopUI2.WPF.csproj", "{4BC74929-71C6-4B5C-9325-491BAC540575}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DesktopUI2.WPF", "DesktopUI2.WPF\DesktopUI2.WPF.csproj", "{4BC74929-71C6-4B5C-9325-491BAC540575}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AvaloniaHwndHost", "AvaloniaHwndHost\AvaloniaHwndHost.csproj", "{421B0919-F7DB-4283-A5A2-009267428776}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -51,6 +53,14 @@ Global
 		{4BC74929-71C6-4B5C-9325-491BAC540575}.Release|Any CPU.Build.0 = Release|Any CPU
 		{4BC74929-71C6-4B5C-9325-491BAC540575}.Release|x86.ActiveCfg = Release|Any CPU
 		{4BC74929-71C6-4B5C-9325-491BAC540575}.Release|x86.Build.0 = Release|Any CPU
+		{421B0919-F7DB-4283-A5A2-009267428776}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{421B0919-F7DB-4283-A5A2-009267428776}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{421B0919-F7DB-4283-A5A2-009267428776}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{421B0919-F7DB-4283-A5A2-009267428776}.Debug|x86.Build.0 = Debug|Any CPU
+		{421B0919-F7DB-4283-A5A2-009267428776}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{421B0919-F7DB-4283-A5A2-009267428776}.Release|Any CPU.Build.0 = Release|Any CPU
+		{421B0919-F7DB-4283-A5A2-009267428776}.Release|x86.ActiveCfg = Release|Any CPU
+		{421B0919-F7DB-4283-A5A2-009267428776}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
